uh, correct-er fretboard notes
