+++ src/types/music.d.ts
@@ -1,0 +1,6 @@
+export interface ScaleSuggestion {
+  name: string;
+  quality: 'perfect' | 'good' | 'possible';
+  notes: string[];
+  intervals: string[];
+}
